reorganize dir
