+++ h2integrate/core/supported_models.py
@@ -1,9 +1,9 @@
 from h2integrate.resource.river import RiverResource
 from h2integrate.core.feedstocks import FeedstockCostModel, FeedstockPerformanceModel
 from h2integrate.transporters.pipe import PipePerformanceModel
-from h2integrate.core.npv_financial import NPVFinancial
 from h2integrate.transporters.cable import CablePerformanceModel
 from h2integrate.converters.steel.steel import SteelPerformanceModel, SteelCostAndFinancialModel
+from h2integrate.finances.npv_financial import NPVFinancial
 from h2integrate.converters.wind.wind_plant import WindPlantCostModel, WindPlantPerformanceModel
 from h2integrate.finances.profast_financial import ProFastComp
 from h2integrate.converters.hopp.hopp_wrapper import HOPPComponent
